Added default navmenu, removed navmenu from ExampleChecker
