--- modules/example_check_module/menu/Style.js
@@ -1,22 +1,0 @@
- var style = {
-  text: {
-    cursor: 'pointer',
-    paddingLeft: '15px',
-    paddingRight: '5px'
-  },
-  statusIcon: {
-    yes: {
-      color: 'green',
-      display: 'initial'
-    },
-    no: {
-      color: 'red',
-      display: 'initial'
-    },
-    unchecked: {
-      display: 'none'
-    }
-  }
-};
-
-module.exports = style;
