+++ src/js/components/core/navigation_menu/NavigationMenu.js
