+++ src/js/components/core/navigation_menu/Style.js
@@ -1,0 +1,28 @@
+ var style = {
+  menuItem: {
+    text: {
+      cursor: 'pointer',
+      paddingLeft: '15px',
+      paddingRight: '5px'
+    },
+    statusIcon: {
+      retained: {
+        color: 'green',
+        display: 'initial'
+      },
+      replaced: {
+        color: 'gold',
+        display: 'initial'
+      },
+      wrong: {
+        color: 'red',
+        display: 'initial'
+      },
+      unchecked: {
+        display: 'none'
+      }
+    }
+  }
+};
+
+module.exports = style;
